+++ kamehouse-groot/public/kame-house-groot/admin/server-manager/index.php
@@ -709,6 +709,9 @@
     </div>
     <br>
   </div>
+  <div class="txt-c-d-kh m-20-d-kh">
+    <a><img class="img-btn-kh server-manager-batcave" src="/kame-house/img/dc/batman-1989-black.png" alt="Batcave" onclick='window.open("/kame-house-batcave/index.html")'/></a>
+  </div>
   <span id="debug-mode-wrapper"></span>
   <script src="/kame-house/js/admin/tomcat-module-status-manager.js"></script>
   <script src="/kame-house-groot/js/admin/kamehouse-shell/tail-log-manager.js"></script>