+++ kamehouse-shell/bin/kamehouse/deploy-kamehouse-dev.sh
@@ -1,20 +1,20 @@
 #!/bin/bash
 
 # Import common functions
-source ${HOME}/programs/kamehouse-shell/bin/common/common-functions.sh
+source ${HOME}/programs/kamehouse-shell/bin/common/functions/common-functions.sh
 if [ "$?" != "0" ]; then
 	echo -e "\033[1;36m$(date +%Y-%m-%d' '%H:%M:%S)\033[0;39m - [\033[1;31mERROR\033[0;39m] - \033[1;31mAn error occurred importing common-functions.sh\033[0;39m"
 	exit 99
 fi
 
 # Import kamehouse functions
-source ${HOME}/programs/kamehouse-shell/bin/common/kamehouse/kamehouse-functions.sh
+source ${HOME}/programs/kamehouse-shell/bin/common/functions/kamehouse/kamehouse-functions.sh
 if [ "$?" != "0" ]; then
   echo -e "\033[1;36m$(date +%Y-%m-%d' '%H:%M:%S)\033[0;39m - [\033[1;31mERROR\033[0;39m] - \033[1;31mAn error occurred importing kamehouse-functions.sh\033[0;39m"
   exit 99
 fi
 
-source ${HOME}/programs/kamehouse-shell/bin/common/kamehouse/docker-functions.sh
+source ${HOME}/programs/kamehouse-shell/bin/common/functions/kamehouse/docker-functions.sh
 if [ "$?" != "0" ]; then
   echo -e "\033[1;36m$(date +%Y-%m-%d' '%H:%M:%S)\033[0;39m - [\033[1;31mERROR\033[0;39m] - \033[1;31mAn error occurred importing docker-functions.sh\033[0;39m"
   exit 99