+++ kamehouse-shell/bin/kamehouse/docker/release/java8-release/java8-release-kamehouse-docker-build.sh
@@ -1,13 +1,13 @@
 #!/bin/bash
 
 # Import common functions
-source ${HOME}/programs/kamehouse-shell/bin/common/common-functions.sh
+source ${HOME}/programs/kamehouse-shell/bin/common/functions/common-functions.sh
 if [ "$?" != "0" ]; then
   echo -e "\033[1;36m$(date +%Y-%m-%d' '%H:%M:%S)\033[0;39m - [\033[1;31mERROR\033[0;39m] - \033[1;31mAn error occurred importing common-functions.sh\033[0;39m"
   exit 99
 fi
 
-source ${HOME}/programs/kamehouse-shell/bin/common/kamehouse/docker/release/java8-release-functions.sh
+source ${HOME}/programs/kamehouse-shell/bin/common/functions/kamehouse/docker/release/java8-release-functions.sh
 if [ "$?" != "0" ]; then
   echo -e "\033[1;36m$(date +%Y-%m-%d' '%H:%M:%S)\033[0;39m - [\033[1;31mERROR\033[0;39m] - \033[1;31mAn error occurred importing java8-release-functions.sh\033[0;39m"
   exit 99