+++ kamehouse-shell/bin/kamehouse/mariadb-csv-kamehouse.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 # Import common functions
-source ${HOME}/programs/kamehouse-shell/bin/common/common-functions.sh
+source ${HOME}/programs/kamehouse-shell/bin/common/functions/common-functions.sh
 if [ "$?" != "0" ]; then
   echo -e "\033[1;36m$(date +%Y-%m-%d' '%H:%M:%S)\033[0;39m - [\033[1;31mERROR\033[0;39m] - \033[1;31mAn error occurred importing common-functions.sh\033[0;39m"
   exit 99
@@ -16,7 +16,7 @@ source ${HOME}/.kamehouse/.shell/.cred
 # Global variables
 LOG_PROCESS_TO_FILE=true
 PATH_CSV=${HOME}/.kamehouse/mariadb/csv
-PATH_SQL=${HOME}/programs/kamehouse-shell/bin/kamehouse/sql/mariadb
+PATH_SQL=${HOME}/programs/kamehouse-shell/sql/mariadb
 NUMBER_OF_BACKUPS=3
 OUT_FILE_BASE=""
 TMP_EXPORT_DIR=/tmp/kamehouse-csv-${USER}