+++ kamehouse-shell/bin/win/bashrc/alias.sh
@@ -1,9 +1,0 @@
-# win aliases
-alias cd.apache2="cd ${HOME}/programs/apache-httpd"
-alias cd.hacking='cd /d/niko9enzo-git/hacking'
-alias cd.learn.java="cd ${HOME}/workspace-eclipse/learn.java"
-alias cd.programming='cd /d/niko9enzo-git/programming'
-alias cd.www="cd ${HOME}/programs/apache-httpd/www/kamehouse-webserver"
-
-alias javac="\"${JAVA_HOME}\"/bin/javac"
-alias javap="\"${JAVA_HOME}\"/bin/javap"