+++ changelog.md
@@ -19,6 +19,7 @@
 - Cleaned up unused kamehouse shell scripts
 - Updated remove special chars script to handle audio files and playlists too
 - Fixed remove special chars script to group multiple special chars into one
+- Reorganized kamehouse shell scripts folders and bashrc functions
 - Fixed docker host ip on linux
 - Fixed error banners background positions on mobile
 - Fixed bugs